Handle null query results in commands

+++ api/src/Command/CleanFiles.php
@@ -47,13 +47,15 @@ class CleanFiles extends Command
         $this->logger->info($this->getName());
         // First we get all files not linked to a user and/or a message
         $c = $this->pdo->query('SELECT f.id, f.content_url FROM file f WHERE NOT EXISTS (SELECT * FROM messages_files mf WHERE mf.file_id = f.id) AND NOT EXISTS (SELECT * FROM user u WHERE u.avatar_id = f.id) AND NOT EXISTS (SELECT * FROM link l WHERE l.preview_id = f.id);');
-        while ($i = $c->fetch()) {
-            if ($input->getOption('verbose') || $input->getOption('only-list')) {
-                echo $this->targetDir.'/'.$i['content_url']."\n";
-            }
-            if (!$input->getOption('only-list')) {
-                $this->pdo->query("DELETE FROM file WHERE id = '".$i['id']."';");
-                unlink($this->targetDir.'/'.$i['content_url']);
+        if ($c !== false) {
+            while ($i = $c->fetch()) {
+                if ($input->getOption('verbose') || $input->getOption('only-list')) {
+                    echo $this->targetDir.'/'.$i['content_url']."\n";
+                }
+                if (!$input->getOption('only-list')) {
+                    $this->pdo->query("DELETE FROM file WHERE id = '".$i['id']."';");
+                    unlink($this->targetDir.'/'.$i['content_url']);
+                }
             }
         }
